PM-14206: State value race condition when subscribing for first time.
When subscribing to the observable, where there have been an update to the state, the current state value and updates values race together due to `merge` operator. The behaviour on which one resolves first is random. Changed the operator to concat to always get sequential observables. Since updates observable holds the newest value, changed this operator to be first, which current value as fallback in case there is no value.